pyproject.toml
[tool.poetry]
name = "softdesk"
version = "0.1.0"
description = ""
authors = ["muumz <useless.muumz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.1"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
